refactor: Auth 명으로 수정
Github issue #64
