+++ src/main/java/org/wedding/adapter/in/web/AuthController.java
@@ -7,7 +7,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import org.wedding.application.service.UserService;
+import org.wedding.adapter.in.web.dto.LoginDTO;
+import org.wedding.adapter.in.web.dto.SignUpDTO;
+import org.wedding.application.port.in.AuthUseCase;
+import org.wedding.common.response.ApiResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -15,24 +18,21 @@ import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.wedding.common.response.ApiResponse;
-import org.wedding.adapter.in.web.dto.LoginDTO;
-import org.wedding.adapter.in.web.dto.SignUpDTO;
 
 @Slf4j
-@Tag(name="User API", description = "User API")
+@Tag(name="Auth API", description = "Auth API")
 @RestController
-@RequestMapping("/api/v1/users")
+@RequestMapping("/api/v1/auth")
 @RequiredArgsConstructor
-public class UserController {
-
-    private final UserService userService;
+public class AuthController {
+    private final AuthUseCase authUseCase;
 
     @PostMapping("/sign-up")
     @Operation(summary = "회원가입", description = "회원가입")
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<ApiResponse<Void>> signUp(@RequestBody @Valid SignUpDTO request) {
-        userService.signUp(request);
+
+        authUseCase.signUp(request);
         ApiResponse<Void> response = ApiResponse.successApiResponse(HttpStatus.CREATED, "회원가입이 완료되었습니다.");
         return new ResponseEntity<>(response, response.status());
     }
@@ -41,9 +41,9 @@ public class UserController {
     @Operation(summary = "로그인", description = "로그인")
     @ResponseStatus(HttpStatus.OK)
     public ResponseEntity<ApiResponse<Void>> login(@RequestBody @Valid LoginDTO request, HttpSession session) {
+
         ApiResponse<Void> response = ApiResponse.successApiResponse(HttpStatus.OK, "로그인 성공했습니다.");
-        userService.login(request, session);
+        authUseCase.login(request, session);
         return new ResponseEntity<>(response, response.status());
     }
-
 }
